Convert notifications to new in memory cache collection
Moves user and global notifications to the new cache collection
scheme that was introduced in #52210. No additional functionality
changes have been made here. This should be a purely mechanical
translation to the new internal caching machinery.
